Added wrappers to normalize rewards

+++ src/test.py
@@ -1,4 +1,5 @@
 import gym
+import numpy as np
 import torch
 
 from actor import Actor
@@ -17,7 +18,7 @@ if __name__ == '__main__':
 
     observation, info = env.reset()
     for _ in range(NUM_STEPS):
-        action = actor(torch.tensor(observation)).sample().detach().numpy()
+        action = actor(torch.tensor(observation, dtype=torch.float32)).sample().detach().numpy()
         observation, reward, terminated, truncated, info = env.step(action)
         if terminated or truncated:
             observation, info = env.reset()
